Moved non-main functions and control data away from we_main.c into library for testability. * Created we_control.h and we_control.c for library * Adapted autotools for we_control * Moved non-main functions and global data to we_control. * Adapted includes in we_main to new placement of functions and global data. * Corrected compilation errors due to change.
